bindEmitter: rewrap properly (and not too much)
Made sure the name of the namespace is available for contexts, so a
handler can be bound to multiple namespaces without barfing. Also
switched to using try-finally for application + cleanup, which may turn
out to be a performance 'no' but feels cleaner. Finally, added
Readable-specific tests to ensure that rewrapping is working properly.