bindEmitter: .removeListener needs unwrapped events
If I'm going to monkeypatch some of the EE behavior, I need to make sure
I catch everything, is the lesson of this story. See the new test case
for details, but the short version is that I'm binding handlers to
namespaces for the duration of an .emit call, but it's possible that a
handler will call .removeListener during that call, and .removeListener
is looking for the unwrapped handler. Several Connect middlewares do
weird things with adding listeners, which triggers an infinite death
spiral of resource leakage.